% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netupdate.R
\name{netupdate}
\alias{netupdate}
\title{Update data based on network configuration}
\usage{
netupdate(...)
}
\arguments{
\item{...}{Stuff that goes in.}
}
\value{
Nothing.
}
\description{
Update data based on network configuration
}
