% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netupdate.R
\name{netupdate}
\alias{netupdate}
\title{Title}
\usage{
netupdate(
  input,
  data,
  wmat,
  scale,
  type = c("standardized", "raw"),
  stdc = 1,
  verbose = TRUE
)
}
\arguments{
\item{verbose}{}
}
\description{
Title
}
